requirements.txt
boto3
PyYAML
python-dotenv
redshift-connector